Problem: not possible to have `MapError` and `AsMap` methods
Solution: drop the `AsMap` - the map can be created via plain function
too and not having a method solves the problem of `comparable`.

IOW keeping `AsMap` will lead to a situation `MapError` will be usable
for `[T, V comparable]` which is unnecessary constraint int this case.
Creating Seq2, which can't be expressed as `map` is perfectly legal and
this mismatch is going to be catch by `AsMap` function anyway.